iroh_n0des

Module caps

Structs

CapSet

A set of capabilities

Enums

Cap

A capability is the capacity to do something. Capabilities are embedded within signed tokens that dictate who created them, and who they apply to. Caps follow the object capability model, where possession of a valid capability token is the canonical source of authorization. This is different from an access control list approach where users authenticate, and their current set of capabilities are stored within a database.

Caps

MetricsCap

NetDiagnosticsCap

RelayCap

TicketsCap

Functions

create_api_token_from_secret_key

Create an rcan token for the api access from an iroh secret key

create_api_token_from_ssh_key

Create an rcan token for the api access.

create_grant_token

Create an rcan token that grants capabilities to a remote endpoint. The local endpoint is the issuer (granter), and the remote endpoint is the audience (grantee).