iroh_n0des

Module protocol

Structs

Auth

Authentication on first request

GetTicket

wire-level request get a ticket by name.

GrantCap

Grant a capability token to the remote endpoint. The remote should store the RCAN and use it when dialing back to authorize its requests.

ListTickets

Wire format for requesting a list of tickets

Ping

Simple ping requests

Pong

Simple ping response

PublishTicket

Publishing a ticket allows n0des to act as a central hub to ferry tickets between endpoints.

PutMetrics

Request to store the given metrics data

PutNetworkDiagnostics

Publishing network diagnostics

RunNetworkDiagnostics

ask this node to run diagnostics & return the result. present even without the net_diagnostics feature flag because the request struct is empty in both cases

TicketData

Signals are opaque data that n0des can ferry between endpoints

UnpublishTicket

wire-level request to remove a ticket. Useful for undos, and any situation where the uptime of the endpoint outlasts the utility of the ticket.

Enums

ClientHostProtocol

Dedicated protocol for cloud-to-endpoint net diagnostics connections.

N0desMessage

Message enum for N0desProtocol

N0desProtocol

NetDiagnosticsMessage

Message enum for ClientHostProtocol

RemoteError

Constants

ALPN

The main ALPN for connecting from the client to the cloud node.

Type Aliases

N0desClient

RemoteResult