iroh_n0des::simulation

Trait AsyncCallback

pub trait AsyncCallback<'a, A1: 'a, A2: 'a, T: 'a>:
    'static
    + Send
    + Sync
    + Fn(&'a mut A1, &'a A2) -> Self::Fut {
    type Fut: Future<Output = T> + Send;
}

Helper trait for async functions.

This is needed because with a simple impl Fn() -> Fut, we can’t express a variadic lifetime bound from the future to the function parameter. impl AsyncFn would allow this, but that doesn’t allow to express a Send bound on the future.

Required Associated Types

type Fut: Future<Output = T> + Send