iroh_n0des
Module protocol

Structs

Auth

Authentication on first request

DeleteTopic

Request that the node joins the given tossip topic

GetTag

Request the name of a blob held by the node

N0desService

Ping

Simple ping requests

Pong

Simple ping response

PutBlob

Request that the node fetches the given blob.

PutMetrics

Request to store the given metrics data

PutTopic

Request that the node joins the given tossip topic

Enums

N0desMessage

N0desProtocol

RemoteError

Constants

ALPN

Type Aliases

N0desClient

ProtoTopicId

RemoteResult